% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats_deconvolution.R
\name{sumstats_deconvolution}
\alias{sumstats_deconvolution}
\title{Format summary statistics and run the deconvolution procedure}
\usage{
sumstats_deconvolution(
  sumstats,
  ld_inverse_dir,
  index,
  parallel = F,
  nbcores = 1
)
}
\arguments{
\item{sumstats}{A data frame containing exactly two columns:
\describe{
\item{variant}{Character vector of variant identifiers in the format
\code{CHR:POS:A1:A2}, based on hg19 genomic coordinates.}
\item{tstat}{Numeric vector of test statistics to be deconvolved.}
}}

\item{ld_inverse_dir}{Path to the directory containing the LD inverse matrices.
These files can be downloaded from Zenodo at \code{zenodo/abcd.fr}.}

\item{index}{A variant index based on hg19 coordinates, used to match the
variants in \code{sumstats} to their corresponding positions in the LD inverse
matrices. This file can be downloaded from Zenodo at \code{zenodo/abcd.fr}.}

\item{parallel}{Indicating whether block-wise computation
should be performed in parallel using \code{parallel::mclapply()}.
Parallel computation might be RAM expensive.}

\item{nbcores}{Number of CPU cores to use when \code{parallel = TRUE}.}
}
\value{
A data frame containing, for each processed variant, the index,
the original test statistic, and the deconvolved test statistic.
}
\description{
Prepares input summary statistics for the deconvolution algorithm
and maps variants to the corresponding entries in the LD inverse matrices.
}
